(Refactor): Fix linting and additional refactoring[sk4122]
