+++ aignostic/tests/model_api/unit/output.txt
@@ -1,0 +1,1 @@
+{"column_names":["response"],"rows":[[{"label":"POSITIVE","score":0.9964361190795898}]]}
